samples/azure_foundry_sdk/requirements.txt
azure-ai-projects>=2.0.0
azure-identity>=1.19.0
